features: add gtag for google analytics. (analytics)
manuals:
improve readme.md and the introduction in the welcome component.

user interfaces:
add a logo to a header. move the location of action component. add loading sign when lazy-loading start. the editor and components are positioned in the tablet in the same way as the mobile. (ui)

change a way to load workspace module to lazy-loading. clean unused module and add gzip option to .htaccess for the compression. (performance)

add tests for the markdwon editor and make some tests more better. (tests)

check that new commit can be submitted. make it not to load vscode loader when a device is mobile. (features)